backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
reportlab==4.2.2